Make location change publisher
