+++ corehq/apps/change_feed/data_sources.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 from corehq.apps.change_feed.exceptions import UnknownDocumentStore
+from corehq.apps.locations.document_store import ReadonlyLocationDocumentStore
 from corehq.apps.sms.document_stores import ReadonlySMSDocumentStore
 from corehq.form_processor.document_stores import (
     ReadonlyFormDocumentStore, ReadonlyCaseDocumentStore, ReadonlyLedgerV2DocumentStore,
@@ -15,6 +16,7 @@ CASE_SQL = 'case-sql'
 SMS = 'sms'
 LEDGER_V2 = 'ledger-v2'
 LEDGER_V1 = 'ledger-v1'
+LOCATION = 'location'
 
 
 def get_document_store(data_source_type, data_source_name, domain):
@@ -36,6 +38,8 @@ def get_document_store(data_source_type, data_source_name, domain):
         return ReadonlyLedgerV2DocumentStore(domain)
     elif data_source_type == LEDGER_V1:
         return LedgerV1DocumentStore(domain)
+    elif data_source_type == LOCATION:
+        return ReadonlyLocationDocumentStore(domain)
     else:
         raise UnknownDocumentStore(
             'getting document stores for backend {} is not supported!'.format(data_source_type)
