+++ corehq/apps/change_feed/topics.py
@@ -19,6 +19,7 @@ LEDGER = 'ledger'
 COMMCARE_USER = 'commcare-user'
 GROUP = 'group'
 WEB_USER = 'web-user'
+LOCATION = 'location'
 
 
 ALL = (
@@ -34,6 +35,7 @@ ALL = (
     SMS,
     WEB_USER,
     APP,
+    LOCATION,
 )
